Use webworkers for running query
Using webworkers to run query so that the browser does not freeze when running queries.
Note: Could not get react + nextjs to play nicely with offscreen canvas workers, so
manually passed in imageData. This is not ideal.

Might be worth tearing out react and using typescript on its own to manipulate the dom.
Not sure if actually desirable