Implement parallel web workers (glitch-effect)
Saving this version, as the resulting effects are pretty interesting.
They are a result of batching the image processing to multiple workers
then appending the end results as they come in, resulting in the image
parts coming out of order

I suspect there is another issue because the final image has some
horizontal slicing instead of just vertical slicing. Also there
are a lot of exceptions.